fraudDetk/fraud_detection/requirements.txt
flask
joblib
numpy
mysql-connector-python
scikit-learn
streamlit
pandas
requests
plotly
scikit-learn